// src/app/interfaces/IBenchmarkRecord.ts
export interface IBenchmarkRecord {
  benchmark_version: number;
  calculation_total: number;
  id: string;
  ifelse: number;
  loops: number;
  math: number;
  php_version: number;
  platform: string;
  server_addr: string;
  server_name: string;
  string: string;
  timestamp: string;
  total: number;
  xdebug: number;
}